get_consensus now returns a list of invariant sites, which main removes
Previously, get_consensus only printed out the list of sites, but this
turned out to mess with numpy.corr_coef, which gives NaN's if it's all
the same. I thus modified the signature of get_consensus to return both
the consensus sequence and the list of sites to remove. I also then
remove them before proceeding.

In practice, these invariant sites seemed to be all gaps in the
alignment anyways. Stupid starting alignment.