Swithed to "production strength" libraries
This code was originally written for the Python Course, which assumed a
limited set of tools going in. If there's a chance this'll be used by
anyone else, this will make the code less fragile and more easily
configurable.

Added in argparse code to take sequence file from the command line

Put code in an if __name__ == "__main__" block

Switched from homebrew fasta parser to BioPython